+++ cmsafvis/NEWS.md
@@ -1,5 +1,10 @@
 # Changelog for cmsafvis package
 
+## 1.2.5
+
+- New attribute for quicklook config file: `logmin`
+- Bug fix for `quicklook` for incorrect plot of grey background in some cases
+
 ## 1.2.4
 
 - Minor update of `quicklook` for colour scale with triangle